Bug 1863256 - Support float values for gecko-profile interval in raptor/talos. r=canaltinova,perftest-reviewers,kshampur DONTBUILD
This patch fixes an issue where we don't accept float values for gecko-profile-interval. It's possible to use float values here so we should support them like in other components.

Differential Revision: https://phabricator.services.mozilla.com/D207859

+++ testing/mozharness/mozharness/mozilla/testing/talos.py
@@ -188,7 +188,7 @@ class Talos(
                 ["--gecko-profile-interval"],
                 {
                     "dest": "gecko_profile_interval",
-                    "type": "int",
+                    "type": "float",
                     "help": "The interval between samples taken by the profiler (milliseconds)",
                 },
             ],
